Fixed format-security error and updated Makefile
Errors in my pull request resulted in these errors:

They didn't happen when I built here. It turns out that the build
on github is using the -Wformat and -Werror=format-security
args during the compile. I added those to the Makefile, so this
error would come up in the future if anyone made it, and then
fixed the bugs exhibiting this problem.