Fix for size_t formatting error on Windows
This isn't related to the code I changed in my pull request, but
it's causing an error in the github build, so here's my attempt at
fixing it. Since the compilation error doesn't happen on my
Raspberry Pi where I'm compiling it, I have no way to know if this
fix works except logic and waiting to see if the update to the
pull request passes the tests.