Fixed dllama-api bug with java's HttpURLConnection
The existing readHttpRequest() method in dllama-api assumed the entire
request came in all in one chunk. While this is true for simple curl
invocations and some c-libraries, it was not true for code using Java's
HttpURLConnection. Prior to this fix, any POST requests coming from
HttpURLConnection (or any other code using chunked transfer encoding)
would have their headers make it through but the body would be completely
dropped, resulting in an empty LLM prompt, causing random replies that had
nothing to do with the query.